+++ packages/client/integrations/jira/JiraClientIntegration.ts
@@ -3,12 +3,15 @@ import type Atmosphere from '../../Atmosphere'
 import JiraSVG from '../../components/JiraSVG'
 import JiraProjectId from '../../shared/gqlIds/JiraProjectId'
 import {jiraIntegrationMeta} from '../../shared/integrations/jiraIntegrationMeta'
+import atlassianLogo from '../../styles/theme/images/graphics/atlassian-gradient.svg'
 import {ExternalLinks} from '../../types/constEnums'
 import AtlassianClientManager from '../../utils/AtlassianClientManager'
+import {describeAtlassianDisconnect} from '../../utils/atlassianScopes'
 import {
   type ClientIntegrationCapabilities,
   ClientIntegrationDefinition,
-  type ConnectParams
+  type ConnectParams,
+  type ProviderLogoAsset
 } from '../platform/ClientIntegrationDefinition'
 
 export class JiraClientIntegration extends ClientIntegrationDefinition {
@@ -16,6 +19,7 @@ export class JiraClientIntegration extends ClientIntegrationDefinition {
   readonly title = jiraIntegrationMeta.title
   readonly description = jiraIntegrationMeta.description
   readonly Icon = JiraSVG
+  readonly logo: ProviderLogoAsset = {src: atlassianLogo}
   readonly capabilities: ClientIntegrationCapabilities = {
     scoping: {
       Panel: lazy(
@@ -25,6 +29,9 @@ export class JiraClientIntegration extends ClientIntegrationDefinition {
           )
       ),
       projectFilterLabel: (filter) => JiraProjectId.split(filter).projectKey
+    },
+    settings: {
+      getDisconnectSubline: describeAtlassianDisconnect
     }
   }
   readonly authorizationHelpUrl = ExternalLinks.INTEGRATIONS_SUPPORT_JIRA_AUTHORIZATION