+++ packages/client/modules/teamDashboard/components/ProviderList/ProviderList.tsx
@@ -4,16 +4,15 @@ import {type PreloadedQuery, usePreloadedQuery} from 'react-relay'
 import {useLocation} from 'react-router'
 import type {ProviderListQuery} from '../../../../__generated__/ProviderListQuery.graphql'
 import SettingsWrapper from '../../../../components/Settings/SettingsWrapper'
+import {
+  getClientIntegration,
+  isRegisteredClientIntegration
+} from '../../../../integrations/platform/registry'
 import {Providers} from '../../../../types/constEnums'
-import {hasConfluenceScopes, hasJiraScopes} from '../../../../utils/atlassianScopes'
-import AtlassianProviderRow from '../ProviderRow/AtlassianProviderRow'
-import AzureDevOpsProviderRow from '../ProviderRow/AzureDevOpsProviderRow'
+import {hasConfluenceScopes} from '../../../../utils/atlassianScopes'
 import ConfluenceProviderRow from '../ProviderRow/ConfluenceProviderRow'
 import GcalProviderRow from '../ProviderRow/GcalProviderRow'
-import GitHubProviderRow from '../ProviderRow/GitHubProviderRow'
-import GitLabProviderRow from '../ProviderRow/GitLabProviderRow'
-import JiraServerProviderRow from '../ProviderRow/JiraServerProviderRow'
-import LinearProviderRow from '../ProviderRow/LinearProviderRow'
+import IntegrationServiceProviderRow from '../ProviderRow/IntegrationServiceProviderRow'
 import MattermostProviderRow from '../ProviderRow/MattermostProviderRow'
 import MSTeamsProviderRow from '../ProviderRow/MSTeamsProviderRow'
 import SlackProviderRow from '../ProviderRow/SlackProviderRow'
@@ -21,54 +20,30 @@ import SlackProviderRow from '../ProviderRow/SlackProviderRow'
 interface Props {
   queryRef: PreloadedQuery<ProviderListQuery>
   teamId: string
-  retry: () => void
 }
 
 const query = graphql`
   query ProviderListQuery($teamId: ID!) {
     viewer {
-      ...AtlassianProviderRow_viewer
       ...ConfluenceProviderRow_viewer
-      ...JiraServerProviderRow_viewer
-      ...GitHubProviderRow_viewer
-      ...GitLabProviderRow_viewer
       ...MattermostProviderRow_viewer
       ...SlackProviderRow_viewer
-      ...AzureDevOpsProviderRow_viewer
       ...MSTeamsProviderRow_viewer
       ...GcalProviderRow_viewer
-      ...LinearProviderRow_viewer
       teamMember(teamId: $teamId) {
+        services {
+          service
+          title
+          isAvailable
+          isConnected
+          ...IntegrationServiceProviderRow_service
+        }
         integrations {
           atlassian {
             accessToken
             scope
           }
-          jiraServer {
-            auth {
-              id
-              isActive
-            }
-            sharedProviders {
-              id
-            }
-          }
           gcal {
-            auth {
-              id
-            }
-          }
-          github {
-            accessToken
-          }
-          gitlab {
-            auth {
-              provider {
-                scope
-              }
-            }
-          }
-          linear {
             auth {
               id
             }
@@ -82,11 +57,6 @@ const query = graphql`
           }
           slack {
             isActive
-          }
-          azureDevOps {
-            auth {
-              accessToken
-            }
           }
           msTeams {
             auth {
@@ -102,7 +72,7 @@ const query = graphql`
 `
 
 const ProviderList = (props: Props) => {
-  const {queryRef, retry, teamId} = props
+  const {queryRef, teamId} = props
   const data = usePreloadedQuery<ProviderListQuery>(query, queryRef)
   const {viewer} = data
   const {hash} = useLocation()
@@ -111,66 +81,52 @@ const ProviderList = (props: Props) => {
     document.getElementById(hash.slice(1))?.scrollIntoView({block: 'start'})
   }, [hash])
   const integrations = viewer.teamMember?.integrations
+  const services = viewer.teamMember?.services ?? []
+
+  const taskIntegrations = services
+    .filter(
+      ({service, isAvailable}) =>
+        isRegisteredClientIntegration(service) &&
+        (isAvailable || !!getClientIntegration(service).contactUs)
+    )
+    .map((integrationService) => ({
+      name: integrationService.title,
+      connected: integrationService.isConnected && integrationService.isAvailable,
+      component: (
+        <IntegrationServiceProviderRow
+          key={integrationService.service}
+          teamId={teamId}
+          serviceRef={integrationService}
+        />
+      )
+    }))
 
   const allIntegrations = [
+    ...taskIntegrations,
     {
-      name: Providers.JIRA_CLOUD_NAME,
-      connected:
-        !!integrations?.atlassian?.accessToken && hasJiraScopes(integrations?.atlassian?.scope),
-      component: (
-        <AtlassianProviderRow key='atlassian' teamId={teamId} retry={retry} viewer={viewer} />
-      )
-    },
-    {
-      name: 'Atlassian Confluence',
+      name: Providers.CONFLUENCE_NAME,
       connected:
         !!integrations?.atlassian?.accessToken &&
         hasConfluenceScopes(integrations?.atlassian?.scope),
       component: <ConfluenceProviderRow key='confluence' teamId={teamId} viewerRef={viewer} />
     },
     {
-      name: 'Jira Data Center',
-      connected:
-        !!integrations?.jiraServer?.auth?.isActive && integrations.jiraServer?.sharedProviders[0],
-      component: <JiraServerProviderRow key='jira' teamId={teamId} viewerRef={viewer} />
-    },
-    {
-      name: 'GitHub',
-      connected: !!integrations?.github?.accessToken,
-      component: <GitHubProviderRow key='github' teamId={teamId} viewer={viewer} />
-    },
-    {
-      name: 'GitLab',
-      connected: !!integrations?.gitlab.auth,
-      component: <GitLabProviderRow key='gitlab' teamId={teamId} viewerRef={viewer} />
-    },
-    {
-      name: 'Linear',
-      connected: !!integrations?.linear?.auth,
-      component: <LinearProviderRow key='linear' teamId={teamId} viewerRef={viewer} />
-    },
-    {
-      name: 'Mattermost',
+      name: Providers.MATTERMOST_NAME,
       connected: !!integrations?.mattermost.auth,
       component: <MattermostProviderRow key='mm' teamId={teamId} viewerRef={viewer} />
     },
     {
-      name: 'Slack',
-      connected: integrations?.slack?.isActive,
+      name: Providers.SLACK_NAME,
+      connected: !!integrations?.slack?.isActive,
       component: <SlackProviderRow key='slack' teamId={teamId} viewer={viewer} />
     },
     {
-      name: 'Azure DevOps',
-      connected: !!integrations?.azureDevOps.auth?.accessToken,
-      component: <AzureDevOpsProviderRow key='azure' teamId={teamId} viewerRef={viewer} />
-    },
-    {
-      name: 'MS Teams',
+      name: Providers.MSTEAMS_NAME,
       connected: !!integrations?.msTeams.auth,
       component: <MSTeamsProviderRow key='teams' teamId={teamId} viewerRef={viewer} />
     },
     {
-      name: 'Gcal Integration',
+      name: Providers.GCAL_NAME,
       connected: !!integrations?.gcal?.auth,
       component: <GcalProviderRow key='gcal' viewerRef={viewer} teamId={teamId} />
     }