+++ CHANGELOG.md
@@ -5,6 +5,28 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## [13.46.0](https://github.com/ParabolInc/parabol/compare/v13.45.2...v13.46.0) (2026-09-18)
+
+
+### Added
+
+* Add trendline to animation ([#13519](https://github.com/ParabolInc/parabol/issues/13519)) ([e069d0e](https://github.com/ParabolInc/parabol/commit/e069d0ef2e4ba4dbf2b20d8cf5ebbd7d0dc0311e))
+* explain the team health question draw in the template editor ([#13516](https://github.com/ParabolInc/parabol/issues/13516)) ([1299aa9](https://github.com/ParabolInc/parabol/commit/1299aa9e87806aa293923885f62495884ba09992))
+
+
+### Changed
+
+* generic ProviderList host for task integrations (WIP) ([#13476](https://github.com/ParabolInc/parabol/issues/13476)) ([c29b543](https://github.com/ParabolInc/parabol/commit/c29b543755f9b2fa03485de61aa5c40d8d3fe5cc))
+
+## [13.45.2](https://github.com/ParabolInc/parabol/compare/v13.45.1...v13.45.2) (2026-09-17)
+
+
+### Fixed
+
+* change template type in search results for team health ([#13510](https://github.com/ParabolInc/parabol/issues/13510)) ([0094ea3](https://github.com/ParabolInc/parabol/commit/0094ea341217e691766c6316ea0644290ad97b41))
+* TaskCard padding ([#13515](https://github.com/ParabolInc/parabol/issues/13515)) ([5360f22](https://github.com/ParabolInc/parabol/commit/5360f22ed8cf158c7b2368c3ce4b64d8ae29d479))
+* team menu selector ([#13514](https://github.com/ParabolInc/parabol/issues/13514)) ([c86a187](https://github.com/ParabolInc/parabol/commit/c86a187296a3f9ff1f7682a1bfba12812e8ae19a))
+
 ## [13.45.1](https://github.com/ParabolInc/parabol/compare/v13.45.0...v13.45.1) (2026-09-16)